Creates custom hooks for mutation request
