+++ features/issues/components/issue-list/issue-list.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
-import { useMutation, useQueryClient } from "react-query";
 import { useDebouncedCallback } from "use-debounce";
-import { Issue, IssueStatus, updateIssue, useIssues } from "@features/issues";
+import { IssueStatus, useIssues } from "@features/issues";
 import { ProjectLanguage, useProjects } from "@features/projects";
 import { IssueRow } from "./issue-row";
 import * as I from "./issue-list.style";
@@ -13,6 +12,7 @@ import {
   optionByLevel,
   optionByStatus,
 } from "@features/issues/api/select-issues-data";
+import { useResolveMutateIssues } from "../../api/use-resolve-mutate-issues";
 
 export function IssueList() {
   const router = useRouter();
@@ -77,61 +77,6 @@ export function IssueList() {
     }
   };
 
-  //this is where it starts
-  const queryClient = useQueryClient();
-  const resolveIssueId = useMutation(updateIssue, {
-    onMutate: async (variables) => {
-      const { toResolveIds, status } = variables;
-      await queryClient.cancelQueries(["issues", page, status, level, project]);
-
-      const previousData = queryClient.getQueryData<{ items: Issue[] }>([
-        "issues",
-        page,
-        status,
-        level,
-        project,
-      ]);
-
-      if (!previousData) return;
-      // Update the cache with the resolved issues
-      const newData = previousData.items.map((issue) => {
-        if (toResolveIds.includes(issue.id)) {
-          return {
-            ...issue,
-            status,
-          };
-        } else {
-          return issue;
-        }
-      }) as Issue[];
-
-      queryClient.setQueryData<{ items: Issue[] }>(
-        ["issues", page, status, level, project],
-        {
-          items: newData,
-        }
-      );
-
-      return { previousData };
-    },
-    onSuccess: (_, variables) => {
-      const { status } = variables;
-      queryClient.invalidateQueries(["issues", page, status, level, project]);
-      queryClient.invalidateQueries("issues");
-    },
-
-    onError: (error) => {
-      console.error(error);
-      queryClient.cancelQueries(["issues", page, status, level, project]);
-    },
-
-    onSettled: () => {
-      queryClient.invalidateQueries(["issues", page, status, level, project]);
-    },
-  });
-
-  //this is where it ends
-
   const page = Number(router.query.page || 1);
   const navigateToPage = (newPage: number) => {
     router.push({
@@ -147,6 +92,7 @@ export function IssueList() {
 
   const issuesPage = useIssues(page, statusParam, levelParam, projectParam);
   const projects = useProjects();
+  const resolveIssueId = useResolveMutateIssues();
 
   if (issuesPage.isLoading) {
     return <LoadingScreen />;
